Move display option logic into its own hook
