+++ src/DashboardUI/src/components/WaterRightsTab.tsx
@@ -12,10 +12,9 @@ import useProgressIndicator from "../hooks/useProgressIndicator";
 import { useDebounceCallback } from "@react-hook/debounce";
 import { useMapErrorAlert, useNoMapResults } from "../hooks/useMapAlert";
 import { PriorityDateRange } from "./PriorityDateRange";
-import { waterRightsProperties, pointSizes, nldi, colorList } from "../config/constants";
+import { waterRightsProperties } from "../config/constants";
 import { useBeneficialUses, useOwnerClassifications, useStates, useWaterSourceTypes } from "../hooks/useSystemQuery";
-import { defaultPointCircleRadius, defaultPointCircleSortKey, flowPointCircleSortKey, volumePointCircleSortKey } from "../config/maps";
-import useLastKnownValue from "../hooks/useLastKnownValue";
+import { mapLayerNames } from "../config/maps";
 import { useRiverBasinOptions } from '../hooks';
 import { getRiverBasinPolygonsByName } from '../accessors/systemAccessor';
 import { useQuery } from 'react-query';
@@ -23,36 +22,17 @@ import { Accordion } from "react-bootstrap";
 import '../App.scss';
 import BeneficialUseSelect from "./BeneficialUseSelect";
 import { BeneficialUseListItem } from "../data-contracts/BeneficialUseListItem";
-import Icon from "@mdi/react";
-import { mdiMapMarker } from "@mdi/js";
 import BootstrapSwitchButton from 'bootstrap-switch-button-react';
 import { FeatureCollection } from "geojson";
 import { DataPoints, Directions } from "../data-contracts/nldi";
 import Select from "react-select";
-import { FilterContext, NldiFilters, WaterRightsFilters, defaultFilters, defaultNldiFilters } from "../FilterProvider";
+import { FilterContext, NldiFilters, defaultFilters, defaultNldiFilters } from "../FilterProvider";
 import { AccordionEventKey } from "react-bootstrap/esm/AccordionContext";
 import useNldiMapPopup from "../hooks/map-popups/useNldiMapPopup";
 import useWaterRightDigestMapPopup from "../hooks/map-popups/useWaterRightDigestMapPopup";
 import NldiTab from "./NldiTab";
 import useNldiMapFiltering from "../hooks/useNldiMapFiltering";
-
-enum MapGrouping {
-  BeneficialUse = "bu",
-  OwnerClassification = "oClass",
-  WaterSourceType = "wsType"
-}
-
-interface WaterRightsDisplayOptions {
-  pointSize: 'd' | 'f' | 'v',
-  mapGrouping: MapGrouping
-}
-
-const mapDataTiers = [
-  'https://api.maptiler.com/tiles/1e068efa-3cd5-4509-a1aa-286c135fc85c/tiles.json?key=IauIQDaqjd29nJc5kJse',
-  'https://api.maptiler.com/tiles/61b00cf0-537e-456c-9d6c-ad1d4a8ec597/tiles.json?key=IauIQDaqjd29nJc5kJse',
-  'https://api.maptiler.com/tiles/c03b0c46-b9c3-4574-8498-cbebca00b871/tiles.json?key=IauIQDaqjd29nJc5kJse',
-  'https://api.maptiler.com/tiles/6d61092a-7c8a-4cd1-8272-d92cf019730c/tiles.json?key=IauIQDaqjd29nJc5kJse'
-]
+import { MapGrouping, useWaterRightsDisplayOptions } from "../hooks/useWaterRightsDisplayOptions";
 
 const mapColorLegendOptions = [
   { value: MapGrouping.BeneficialUse, label: 'Beneficial Use' },
@@ -60,20 +40,16 @@ const mapColorLegendOptions = [
   { value: MapGrouping.WaterSourceType, label: 'Water Source Type' }
 ];
 
-const waterRightsPointsLayer = 'waterRightsPoints';
-const waterRightsPolygonsLayer = 'waterRightsPolygons';
-const waterRightsRiverBasinLayer = 'river-basins';
-const nldiLayer = ['nldi-flowlines', 'nldi-usgs-location', 'nldi-usgs-points'];
+const nldiLayer = [
+  mapLayerNames.nldiFlowlinesLayer,
+  mapLayerNames.nldiUsgsLocationLayer,
+  mapLayerNames.nldiUsgsPointsLayer
+];
 
 const allWaterRightsLayers = [
-  waterRightsPointsLayer,
-  waterRightsPolygonsLayer
+  mapLayerNames.waterRightsPointsLayer,
+  mapLayerNames.waterRightsPolygonsLayer
 ]
-
-const defaultDisplayOptions: WaterRightsDisplayOptions = {
-  pointSize: 'd',
-  mapGrouping: MapGrouping.BeneficialUse
-}
 
 const exemptMapping = new Map<boolean | undefined, '' | '0' | '1'>([
   [undefined, ''],
@@ -113,41 +89,19 @@ function WaterRightsTab() {
     }
     return await getRiverBasinPolygonsByName(filters?.riverBasinNames ?? []);
   }, { keepPreviousData: true });
-  const [displayOptions, setDisplayOptions] = useState<WaterRightsDisplayOptions>(getUrlParam<WaterRightsDisplayOptions>("wrd") ?? defaultDisplayOptions);
-
-  const mapGrouping = useMemo(() => {
-    let colorIndex = 0;
-    let colorMapping: { key: string, color: string }[];
-    switch (displayOptions.mapGrouping) {
-      case MapGrouping.BeneficialUse:
-        colorMapping = allBeneficialUses?.map(a => ({ key: a.beneficialUseName, color: colorList[colorIndex++ % colorList.length] })) ?? []
-        break;
-      case MapGrouping.OwnerClassification:
-        colorMapping = allOwnerClassifications?.map(a => ({ key: a, color: colorList[colorIndex++ % colorList.length] })) ?? []
-        break;
-      case MapGrouping.WaterSourceType:
-        colorMapping = allWaterSourceTypes?.map(a => ({ key: a, color: colorList[colorIndex++ % colorList.length] })) ?? []
-        break;
-    }
-    return { property: displayOptions.mapGrouping as string, colorMapping }
-  }, [displayOptions.mapGrouping, allBeneficialUses, allWaterSourceTypes, allOwnerClassifications])
 
   const {
-    setLegend,
     setLayerFilters: setMapLayerFilters,
     setVisibleLayers,
-    renderedFeatures,
     polylines,
     setPolylines,
-    setLayerCircleColors,
-    setLayerFillColors,
-    setVectorUrl,
     geoJsonData,
     setGeoJsonData
   } = useContext(MapContext);
 
   const [isNldiMapActive, setNldiMapStatus] = useState<boolean>(getUrlParam("nldiActive") ?? false);
   const [activeKeys, setActiveKeys] = useState<AccordionEventKey>(isNldiMapActive ? ["nldi"] : ["colorSizeTools", "siteSelectionFilters"]);
+  const {displayOptions, setMapGrouping, setPointSize, resetDisplayOptions} = useWaterRightsDisplayOptions(filters, allBeneficialUses, allOwnerClassifications, allWaterSourceTypes);
 
   useEffect(() => {
     for (let element of filters.polyline) {
@@ -161,121 +115,6 @@ function WaterRightsTab() {
       polyline: polylines
     }))
   }, [setFilters, polylines])
-
-  useEffect(() => {
-    let params = (new URL(document.location.href)).searchParams;
-    let tier = parseInt(params.get("tier") ?? "");
-    if (!isNaN(tier) && tier >= 0 && tier < mapDataTiers.length) {
-      setVectorUrl('allocation-sites_1', mapDataTiers[tier])
-    }
-  }, [setVectorUrl, isNldiMapActive])
-
-  const renderedMapGroupings = useMemo(() => {
-    const tryParseJsonArray = (value: any) => {
-      try {
-        return JSON.parse(value ?? "[]");
-      } catch (e) {
-        return [];
-      }
-    }
-    let colorMappings = [...mapGrouping.colorMapping];
-    if (mapGrouping.property === MapGrouping.BeneficialUse as string && filters.beneficialUses && filters.beneficialUses?.length > 0) {
-      colorMappings = colorMappings.filter(a => filters.beneficialUses?.some(b => b.beneficialUseName === a.key));
-    }
-    if (mapGrouping.property === MapGrouping.WaterSourceType as string && filters.waterSourceTypes && filters.waterSourceTypes.length > 0) {
-      colorMappings = colorMappings.filter(a => filters.waterSourceTypes?.some(b => b === a.key));
-    }
-    if (mapGrouping.property === MapGrouping.OwnerClassification as string && filters.ownerClassifications && filters.ownerClassifications.length > 0) {
-      colorMappings = colorMappings.filter(a => filters.ownerClassifications?.some(b => b === a.key));
-    }
-    colorMappings = colorMappings.filter(a => renderedFeatures.some(b => b.properties && tryParseJsonArray(b.properties[mapGrouping.property]).some((c: string) => c === a.key)));
-
-    return {
-      property: mapGrouping.property,
-      colorMapping: colorMappings
-    }
-  }, [mapGrouping, renderedFeatures, filters.beneficialUses, filters.waterSourceTypes, filters.ownerClassifications])
-
-  useEffect(() => {
-    let colorArray: any;
-    if (mapGrouping.colorMapping.length > 0) {
-      colorArray = ["case"];
-      renderedMapGroupings.colorMapping.forEach(a => {
-        colorArray.push(["in", a.key, ["get", mapGrouping.property]]);
-        colorArray.push(a.color)
-      })
-      mapGrouping.colorMapping.forEach(a => {
-        colorArray.push(["in", a.key, ["get", mapGrouping.property]]);
-        colorArray.push(a.color)
-      })
-      colorArray.push("#000000");
-    } else {
-      colorArray = "#000000"
-    }
-
-    setLayerCircleColors({
-      layer: waterRightsPointsLayer,
-      circleColor: colorArray
-    })
-    setLayerFillColors({
-      layer: waterRightsPolygonsLayer,
-      fillColor: colorArray
-    })
-  }, [setLayerCircleColors, setLayerFillColors, mapGrouping, renderedMapGroupings])
-
-  useEffect(() => {
-    if (renderedMapGroupings.colorMapping.length === 0 && !isNldiMapActive) {
-      setLegend(null);
-    } else {
-      setLegend(
-        <>
-          {
-            renderedMapGroupings.colorMapping.map(layer => {
-              return (
-                <div key={layer.key} className="legend-item">
-                  <span className="legend-circle" style={{ "backgroundColor": layer.color }}></span>
-                  {layer.key}
-                </div>
-              )
-            })
-          }
-          {isNldiMapActive &&
-            <div className="legend-nldi">
-              <div className="legend-item">
-                <span>
-                  <Icon path={mdiMapMarker} size="14px" style={{ color: nldi.colors.mapMarker }} />
-                </span>
-                Starting Point of Interest
-              </div>
-              <div className="legend-item">
-                <span className="legend-flowline">
-                  <span className="legend-flowline legend-flowline-main" style={{ backgroundColor: nldi.colors.mainstem }} />
-                </span>
-                Mainstem
-              </div>
-              <div className="legend-item">
-                <span>
-                  <span className="legend-flowline legend-flowline-tributary" style={{ backgroundColor: nldi.colors.tributaries }} />
-                </span>
-                Tributaries
-              </div>
-              {!!((filters.nldiFilterData?.dataPoints ?? DataPoints.None) & DataPoints.Usgs) &&
-                <div className="legend-item">
-                  <span className="legend-circle" style={{ "backgroundColor": nldi.colors.usgs }}></span>
-                  USGS NWIS Sites
-                </div>
-              }
-              {!!((filters.nldiFilterData?.dataPoints ?? DataPoints.None) & DataPoints.Epa) &&
-                <div className="legend-item">
-                  <span className="legend-circle" style={{ "backgroundColor": nldi.colors.epa }}></span>
-                  EPA Water Quality Portal<br /> Sites OSM Standard
-                </div>
-              }
-            </div>
-          }
-        </>);
-    }
-  }, [setLegend, renderedMapGroupings, isNldiMapActive, filters.nldiFilterData?.dataPoints])
 
   const [allocationOwnerValue, setAllocationOwnerValue] = useState(filters.allocationOwner ?? "");
   const nldiWadeSiteIds = useMemo(() => {
@@ -311,29 +150,6 @@ function WaterRightsTab() {
     });
   }, [nldiWadeSiteIds, setNldiIds]);
 
-  useEffect(() => {
-    if (deepEqual(filters, defaultFilters)) {
-      setUrlParam("wr", undefined);
-    } else {
-      setUrlParam("wr", filters);
-    }
-  }, [filters, setUrlParam])
-
-  useEffect(() => {
-    if (deepEqual(displayOptions, defaultDisplayOptions)) {
-      setUrlParam("wrd", undefined);
-    } else {
-      setUrlParam("wrd", displayOptions);
-    }
-  }, [displayOptions, setUrlParam])
-
-  const handleMapGroupingChange = (mapGroupingOption: MapGrouping | undefined) => {
-    setDisplayOptions(s => ({
-      ...s,
-      mapGrouping: mapGroupingOption ? mapGroupingOption : MapGrouping.BeneficialUse
-    }));
-  }
-
   const handlePodPouChange = (e: ChangeEvent<HTMLInputElement>) => {
     setFilters(s => ({
       ...s,
@@ -354,11 +170,12 @@ function WaterRightsTab() {
     }));
   }
 
+  const handleMapGroupingChange = (mapGroupingOption: MapGrouping | undefined) => {
+    setMapGrouping(mapGroupingOption);
+  }
+
   const handlePointSizeChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setDisplayOptions(s => ({
-      ...s,
-      pointSize: e.target.value === "f" ? "f" : e.target.value === "v" ? "v" : "d"
-    }));
+    setPointSize(e.target.value);
   }
 
   const handleBeneficialUseChange = (selectedOptions: BeneficialUseListItem[]) => {
@@ -437,14 +254,14 @@ function WaterRightsTab() {
 
   useEffect(() => {
     let visible = [...allWaterRightsLayers];
-    if ((filters.riverBasinNames?.length ?? 0) > 0) visible.push(waterRightsRiverBasinLayer);
+    if ((filters.riverBasinNames?.length ?? 0) > 0) visible.push(mapLayerNames.riverBasinsLayer);
     if (isNldiMapActive === true) visible.push(...nldiLayer);
 
     setVisibleLayers(visible);
   }, [filters.riverBasinNames, setVisibleLayers, isNldiMapActive])
 
   useEffect(() => {
-    setGeoJsonData("river-basins", riverBasinPolygons ?? {
+    setGeoJsonData(mapLayerNames.riverBasinsLayer, riverBasinPolygons ?? {
       "type": "FeatureCollection",
       "features": []
     });
@@ -547,7 +364,7 @@ function WaterRightsTab() {
 
   const clearMapFilters = () => {
     setFilters(defaultFilters);
-    setDisplayOptions({ ...defaultDisplayOptions });
+    resetDisplayOptions();
     setAllocationOwnerValue("");
     polylines.forEach(a => {
       setPolylines(a.identifier, null);
@@ -560,7 +377,6 @@ function WaterRightsTab() {
   const [isNldiDataFetching] = useNldiMapFiltering(filters.nldiFilterData);
   useNldiMapPopup();
   useWaterRightDigestMapPopup();
-  useWaterRightMapPointScaling(displayOptions.pointSize, filters);
 
   const isLoading = useMemo(() => {
     return isAllBeneficialUsesLoading || isAllWaterSourceTypesLoading || isAllOwnerClassificationsLoading || isAllStatesLoading || isRiverBasinOptionsLoading;
@@ -783,114 +599,3 @@ function WaterRightsTab() {
 }
 
 export default WaterRightsTab;
-
-const flowPointCircleRadiusFlow = ["coalesce", ["get", waterRightsProperties.maxFlowRate as string], 0];
-const volumePointCircleRadiusFlow = ["coalesce", ["get", waterRightsProperties.maxVolume as string], 0];
-function useWaterRightMapPointScaling(pointSize: "d" | "f" | "v", filters: WaterRightsFilters) {
-  const [minVolume, lastKnownMinVolume, setMinVolume] = useLastKnownValue(0);
-  const [maxVolume, lastKnownMaxVolume, setMaxVolume] = useLastKnownValue(100);
-  const [minFlow, lastKnownMinFlow, setMinFlow] = useLastKnownValue(0);
-  const [maxFlow, lastKnownMaxFlow, setMaxFlow] = useLastKnownValue(100);
-  const {
-    renderedFeatures,
-    setLayerCircleRadii,
-    setLayerCircleSortKeys
-  } = useContext(MapContext);
-
-  const pointScaleTimeDelay = 1000;
-  useEffect(() => {
-    //we delay these to give the map time to get new rendered features after the filters change
-    //check this if there are issues setting point sizes
-    setTimeout(() => setMinFlow(undefined), pointScaleTimeDelay);
-    setTimeout(() => setMaxFlow(undefined), pointScaleTimeDelay);
-    setTimeout(() => setMinVolume(undefined), pointScaleTimeDelay);
-    setTimeout(() => setMaxVolume(undefined), pointScaleTimeDelay);
-  }, [filters, setMinFlow, setMaxFlow, setMinVolume, setMaxVolume]);
-
-  const flowVolumeMinMax = useMemo(() => {
-    if (renderedFeatures.length === 0) {
-      return { minFlow: 0, maxFlow: 0, minVolume: 0, maxVolume: 0 }
-    }
-    const values = renderedFeatures
-      .map(a => ({
-        flow: (a.properties?.[waterRightsProperties.maxFlowRate as string]) ?? 0,
-        volume: (a.properties?.[waterRightsProperties.maxVolume as string]) ?? 0
-      }));
-    return values
-      .reduce((prev, curr) => (
-        {
-          minFlow: prev.minFlow === undefined || curr.flow < prev.minFlow ? curr.flow : prev.minFlow,
-          maxFlow: prev.maxFlow === undefined || curr.flow > prev.maxFlow ? curr.flow : prev.maxFlow,
-          minVolume: prev.minVolume === undefined || curr.volume < prev.minVolume ? curr.volume : prev.minVolume,
-          maxVolume: prev.maxVolume === undefined || curr.volume > prev.maxVolume ? curr.volume : prev.maxVolume
-        }
-      ), { minFlow: values[0].flow, maxFlow: values[0].flow, minVolume: values[0].volume, maxVolume: values[0].volume })
-  }, [renderedFeatures])
-
-  useEffect(() => {
-    if (minFlow === undefined || flowVolumeMinMax.minFlow < minFlow) {
-      setMinFlow(flowVolumeMinMax.minFlow)
-    }
-  }, [minFlow, flowVolumeMinMax.minFlow, setMinFlow])
-
-  useEffect(() => {
-    if (maxFlow === undefined || flowVolumeMinMax.maxFlow > maxFlow) {
-      setMaxFlow(flowVolumeMinMax.maxFlow)
-    }
-  }, [maxFlow, flowVolumeMinMax.maxFlow, setMaxFlow])
-
-  useEffect(() => {
-    if (minVolume === undefined || flowVolumeMinMax.minVolume < minVolume) {
-      setMinVolume(flowVolumeMinMax.minVolume)
-    }
-  }, [minVolume, flowVolumeMinMax.minVolume, setMinVolume])
-
-  useEffect(() => {
-    if (maxVolume === undefined || flowVolumeMinMax.maxVolume > maxVolume) {
-      setMaxVolume(flowVolumeMinMax.maxVolume)
-    }
-  }, [maxVolume, flowVolumeMinMax.maxVolume, setMaxVolume])
-
-  const min = useMemo(() => {
-    return pointSize === "f" ? filters.minFlow ?? lastKnownMinFlow : filters.minVolume ?? lastKnownMinVolume;
-  }, [pointSize, lastKnownMinFlow, lastKnownMinVolume, filters.minFlow, filters.minVolume])
-
-  const max = useMemo(() => {
-    return pointSize === "f" ? filters.maxFlow ?? lastKnownMaxFlow : filters.maxVolume ?? lastKnownMaxVolume;
-  }, [pointSize, filters.maxFlow, filters.maxVolume, lastKnownMaxFlow, lastKnownMaxVolume])
-
-  const scaleProperty = useMemo(() => {
-    return pointSize === "f" ? flowPointCircleRadiusFlow : volumePointCircleRadiusFlow;
-  }, [pointSize])
-
-  useEffect(() => {
-    if (pointSize === "d") {
-      setLayerCircleRadii({ layer: waterRightsPointsLayer, circleRadius: defaultPointCircleRadius })
-    } else {
-      if (min === max) {
-        setLayerCircleRadii({ layer: waterRightsPointsLayer, circleRadius: defaultPointCircleRadius })
-      }
-      const valueScaleFactorMinToMax = pointSizes.maxScaleFactorForSizedPoints;//the biggest point will be x times bigger than the smallest
-      const zoomScaleFactorMinToMax = pointSizes.maxPointSize / pointSizes.minPointSize;
-      const valueRange = max - min;
-      const minSizeAtMinimumZoom = pointSizes.minPointSize;
-      const maxSizeAtMinimumZoom = pointSizes.minPointSize * valueScaleFactorMinToMax;
-      const sizeRange = maxSizeAtMinimumZoom - minSizeAtMinimumZoom;
-      //(((vol-min)/(max-min))*(maxPointSize-minPointSize))+minPointSize === this point's size at minimum zoom
-      const minZoomValue = ["min", ["+", ["*", ["/", ["-", scaleProperty, min], valueRange], sizeRange], minSizeAtMinimumZoom], maxSizeAtMinimumZoom]
-      //size at minimum * zoomScaleFactorMinToMax === this point's size at maximum zoom
-      const maxZoomValue = ["*", minZoomValue, zoomScaleFactorMinToMax]
-      setLayerCircleRadii({ layer: waterRightsPointsLayer, circleRadius: ["interpolate", ["linear"], ["zoom"], pointSizes.minPointSizeZoomLevel, minZoomValue, pointSizes.maxPointSizeZoomLevel, maxZoomValue] })
-    }
-  }, [pointSize, scaleProperty, min, max, setLayerCircleRadii])
-
-  useEffect(() => {
-    if (pointSize === "f") {
-      setLayerCircleSortKeys({ layer: waterRightsPointsLayer, circleSortKey: flowPointCircleSortKey })
-    } else if (pointSize === "v") {
-      setLayerCircleSortKeys({ layer: waterRightsPointsLayer, circleSortKey: volumePointCircleSortKey })
-    } else {
-      setLayerCircleSortKeys({ layer: waterRightsPointsLayer, circleSortKey: defaultPointCircleSortKey })
-    }
-  }, [pointSize, setLayerCircleSortKeys])
-}
